Large change regarding Attributes and Pipe component
This commit contains many changes that were hard to fix separate from
one another. The notebooks are currently broken and will be fixed up
afterwards.

- The values for Attributes() should now be passed in when creating a
new object. Failing to do so results in a TypeError.
- Different type of attributes exist, to discriminate between those that
need to be set by the user, and those that the software sets. In the
future traitlets should likely be used.
- Abstract component attributes are moved into subclasses.
- Pipe component has been extended.
- test was rewritten to use pytest.